My gifos con favoritos y recorder funcionando
