+++ index.html
@@ -15,7 +15,7 @@
             <img src="./assets/gifOF_logo.png" alt="logo" class="day">
             <img src="./assets/gifOF_logo_dark.png" alt="logo" class="night">
             <nav>
-                <div class="button colored">Crear Guifos</div>
+                <div class="button colored" id="crearGifo">Crear Guifos</div>
                 <div class="theme-button-container">
                     <div class="theme-dopdown" id="theme-dropdown">
                         <div class="button colored">Elegir tema</div>
@@ -28,7 +28,7 @@
                         <div id="night-button" class="button sailor-night"><span>S</span>ailor Night</div>
                     </div>
                 </div>
-                <a href="./misGifos.html"><span>M</span>is Guifos</a>
+                <a href="./misGifos.html#misGifos" id="misGifos"><span>M</span>is Guifos</a>
             </nav>
         </div>
         <div class="search-container">
@@ -213,5 +213,6 @@
     </main>
 
     <script src="./scripts/script.js" type="module"></script>
+    <script src="./scripts/home.js" type="module"></script>
 </body>
 </html>
