+++ misGifos.html
@@ -13,12 +13,16 @@
         <div class="top-bar">¡Bienvenidos/as a Guifos.com! ——————Donde los gifs están.////// Número de visitas: <span id="numVisits">12.765.803</span>
         </div>
         <div class="main-header">
+            <a href="./index.html" id="backArrow">
+                <img src="./assets/arrow.svg" alt="arrow" class="day">
+            </a>
             <a href="./index.html">
                 <img src="./assets/gifOF_logo.png" alt="logo" class="day">
                 <img src="./assets/gifOF_logo_dark.png" alt="logo" class="night">
             </a>
-            <nav>
-                <div class="button colored">Crear Guifos</div>
+            
+            <nav id="nav">
+                <div class="button colored" id="crearGifo">Crear Guifos</div>
                 <div class="theme-button-container">
                     <div class="theme-dopdown" id="theme-dropdown">
                         <div class="button colored">Elegir tema</div>
@@ -31,14 +35,14 @@
                         <div id="night-button" class="button sailor-night"><span>S</span>ailor Night</div>
                     </div>
                 </div>
-                <a><span>M</span>is Guifos</a>
+                <a href="./misGifos.html#misGifos" id="gifosAnchor"><span>M</span>is Guifos</a>
             </nav>
         </div>
     </header>
     <main>
 
         <section id="creadorGifo">
-            <div class="creator-container">
+            <div class="creator-container" id="creator-container">
                 <div class="top">Crear Guifos</div>
                 <div class="text-container">
                     <img src="./assets/window-img.png" alt="window icon">
@@ -55,15 +59,15 @@
                     </div>
                 </div>
                 <div class="button-container">
-                    <div class="button pale">Cancelar</div>
-                    <div class="button colored">Capturar</div>
+                    <div class="button pale cancel" >Cancelar</div>
+                    <div class="button colored" id="capture">Capturar</div>
                 </div>
             </div>
 
-            <div class="capture-container">
+            <div class="capture-container" id="capture-container">
                 <div class="top">Un Chequeo Antes de Empezar</div>
                 <div class="image-container">
-                    <video src="">
+                    <video id="video">
 
                     </video>
                 </div>
@@ -71,14 +75,14 @@
                     <div class="button colored"><img src="./assets/camera.svg" alt="camara"></div>
                     <div class="button colored" >Capturar</div>
                 </div>
-                <div class="button-container recording" id="RecordingButtons">
+                <div class="button-container" id="recordingButtons">
                     <div class="timer">00.00.03.06</div>
                     <div>
-                        <div class="button colored"><img src="./assets/recording.svg" alt="recording"></div>
-                        <div class="button colored" >Listo</div>
+                        <div class="recordingButton button colored"><img src="./assets/recording.svg" alt="recording"></div>
+                        <div class="recordingButton button colored">Listo</div>
                     </div>
                 </div>
-                <div class="button-container preview" id="RecordingButtons">
+                <div class="button-container" id="previewButtons">
                     <div>
                         <div class="timer">00.00.03.06</div>
                         <div class="button colored"><img src="./assets/forward.svg" alt=""></div>
@@ -91,7 +95,7 @@
                     </div>
                 </div>
              </div>
-             <div class="uploading-container">
+             <div class="uploading-container" id="uploading-container">
                 <div class="top">Subiendo Guifo</div>
                 <div class="image-container">
                     <img src="./assets/globe_img.png" alt="">
@@ -100,7 +104,7 @@
                     <p>Tiempo restante: <strike>38 años</strike> algunos minutos</p>
                 </div>
                 <div class="button-container">
-                    <div class="button pale">Cancelar</div>
+                    <div class="button pale cancel">Cancelar</div>
                 </div>
             </div>
 
@@ -133,8 +137,12 @@
 
         </section>
 
+        
+
     </main>
 
     <script src="./scripts/script.js" type="module"></script>
+    <script src="./scripts/myGifos.js" type="module"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/RecordRTC/5.5.8/RecordRTC.js" type="text/javascript"></script>
 </body>
 </html>
